fix: other fixes
